Fix bug requiring client-side certificate
We're making a wrong check here.
```erlang
    if CertFile /= [] orelse Command == ?SET_CERTIFICATE_FILE_ACCEPT ->
```
meant that in order to proceed, if no certificate file was provided,
then at least the command should be an accept one. Which is not correct.
In TLS we require the server, the one who does the accept call, to
provide a certificate, while the client is not usually required so. It
should instead be required for the server.