Test client certfile availability when corresponds
If the server does `get_peer_certificate`, he'll get it if the client
configured so, and it would get an error if the client didn't. We want
to be sure that the client did indeed that when we called `tcp_to_tls`
with a certificate, the server can see it.