From a gen_server init to the on_load directive
Before OTP 19.0, if the on_load directive fails, any previously loaded
code would become old, essentially leaving the system without any
working and reachable instance of the module. This is not true anymore
since OTP 19.0, so loading the nifs within the on_load directive is
today the cleanest and best code pattern.

It seems indeed very convoluted to have an entire application and
supervision tree with a worker gen_server just to load the nifs. Also,
one of the `handle_info` heads of the gen_server is even a remnant of
the times when `fast_tls` was implemented as a port driver instead of as
nifs, which is a high sign this needed some modernisation.