Introduce 'fips_mode' environment variable
When fips_mode is set to true/false, FIPS mode is enabled/disabled
on NIF load. Note that when the variable is not set, nothing happens.
The variable name and behaviour was chosen to be consistent with
crypto library (https://www.erlang.org/doc/apps/crypto/fips.html).

Also fast_tls:set_fips_mode/1 is introduced which can be used
to explicitly enable/disable FIPS and fast_tls:get_fips_mode/0
is introduced to check whether FIPS mode is enabled or not.